+++ Lab2/Gold.cs
@@ -4,6 +4,6 @@ public class Gold : Customer
 {
     public Gold(string name, string password) : base(name, password)
     {
-        Rabatt = (decimal)0.85;
+        Rabatt = 0.85;
     }
 }
